feat: Add support for forwarding requests to the leader
This commit adds support for forwarding requests to the leader in the `KVStore` struct. The `write` method now checks if there is a current leader and forwards the request to it if necessary. If there is no leader, an error is thrown.

The commit also includes changes to the `raft.proto` file, adding new RPC methods for preparing and joining a cluster.

In addition, this commit introduces a new `NetworkFactory` struct that implements the `RaftNetworkFactory` trait. This factory is responsible for creating network connections to other nodes in the cluster.