feat: Add join request forwarding to leader
This commit adds the functionality to forward a join request to the current leader if the node receiving the request is not the leader. If there is no leader, an error message is returned. The code changes include extracting metrics from the Raft object, checking if the current node is the leader, and forwarding the request using gRPC client calls.